Extract shared types to remove duplication
